Game: set up canvas; centered body elements; set up background color code
I chose to make my game on a 1280x720 canvas. Not completely set on that yet, since I want the game to also be accessible on mobile. But I like the idea of it being landscape and higher resolution so that the player can really absorb the experience of the memory.

One thing I can probably do for mobile is prompt the player to rotate their phone when it's in portrait mode so that I can code less and maintain the landscape mode.