Created json file containing placeholder memories; Game: created 2 preview video rectangles
I planned on creating my text and non-interactive elements in HTML, but decided against it since it caused confusion with positions when I had a p5 canvas element as well. So now, all my game elements are created in p5.

I created a json file to store my memory names, and I thought I could try linking each memory to a different file stored in the Memories folder. Still need to test if this will work or not.

---

To do:
- Integrate memory names into preview videos
- Play a test memory