Game: created allMemories.js that will store all code relating to memories; added title to preview video randomly picked from memories.json list; cue memory when clicked on the preview video
Right now what I got going on is when the user clicks on the preview video, it opens up a full-screen memory video.

I tried thinking of a way to link the memory name (what will be displayed on the thumbnail of the preview video) and cueing the memory to play. My solution was that for each json element, there will be a name property (what the user will see, nicely written with spaces) as well as a file name property (which will be linked to the memory state and the file name containing the memory objects).

No memories have been coded yet.