Game: changed look of interface; moved buttons to Buttons folder; created DeleteButton.js and KeepButton.js
Still deciding on what to do with the look of the interface. I started adding in cutesy colors and rounded rectangles for increased friendliness.

The next thing I want to do is create my first prototype where I have a quickly-made pending memory and current memory to play. I want to check if this layout makes sense for users.