Game: Created dialogs.json to manage list of dialogs; created a dialog box to store dialogs of memories; placed red and green circle in MemoryTest.js to test abstracted shapes in memories; added `memory` state; Journal: added new thoughts on how the buttons and memory sections would change names
For dialog boxes: I decided that I will attribute a different color to each speaker so that the user associates the color to who is talking. On the corner of the dialog box, it will also feature the speaker's name.

I made the area of the dialog box big so the user can easily click on it to switch to the next dialog.