Intro: started prepping intro scene; added Memorywash app thumbnaill; added Download button that changes cursor on hover and will download app when clicked
I started setting up the main elements needed for the intro: a thumbnail that offers a sneak peak of the app and a "Download" button. I kept the color palette bright and bubbly --- after all, Memorywash is all about preserving happy memories.

If there is time, I will add a 'Terms & Conditions' section. outlining the risks of downloading the Memorywash app.