Game: made card decorations draggable onto card for MemoryFathersDay
Completed another memory! Woo!

This one is a creative card decoration activity inspired by Cooking Mama's pizza decoration sessions on the Nintendo. The draggable function was a pain to implement with all the card decorations I have, but I got it to work (best part).

At the beginning, I was trying to figure out whether I should just use jQuery UI for this memory to implement the draggable functions. However, if I went with this method, other issues would arise, such as the inability for the user to take a screenshot of the canvas after their card is decorated, and the difficulty of limiting movement of the draggable items within the card found inside the p5 canvas. Anyways, at the end, I think it was the right decision to use Daniel Shiffman's draggable function that can be used inside p5.