Game: added caption that user can input to InstaPicMemory; added flashy hearts and updated numLikes after user posted picture of salad
Really adore how the floating heart emojis look! I played around with the image's tint and it gave a sparkly pink effect on the hearts.

I used some HTML to code the "Write a caption" text field. It was a long and tedious experience as it made me reflect on how to connect HTML elements with p5. The end result is that the user can enter their own caption for the image and then post it and receive an explosion of likes! Hooray for validation from social media!