MemoryMakingBao: added ability to roll on dough with rolling pin and increase the dough's size; MemoryPlayingOnPhone: tweaked facial features positions
Making bao: I made the rolling pin follow the mouse's y position. Constraining the position was a good way to make sure that the player rolls the pin on top of the dough. I dropped the idea of rotating the dough after every roll (skipping some steps while cooking won't hurt, right? ... says the girl who burns food).

I also mapped the increase rate of the dough with the distance that the pin is rolled to ensure that the player does sufficient labor to increase the dough's size.