MemoryMakingBao: updated all images; fixed audio; Memories: added BackgroundImage.js class to simplify adding background images
I finally got a chance to update the bao images. I wonder how players feel with more polished drawings. I wonder if it evokes some of their own memories, or if the table decorations get in the way of understanding the action. I will need to playtest this to find out.