Game: *MAJOR UPDATE* updated Memorywash interface
I made the "Memory Storage Capacity" bar very simple, with a red box highlighting the excess memory count that decreases as the user selects a memory to delete. I also made the background blue because it aligned with the washing theme and gave it a false calming atmosphere, kinda eery?